Refactored existing resources
