+++ app/Http/Controllers/Tournament/GetCompleteTournamentController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers\Tournament;
 
 use App\Http\Controllers\BasicController;
-use App\Http\Resources\CompleteTournamentResource;
+use App\Http\Resources\Tournament\TournamentResource;
 use App\Models\Tournament;
 use Illuminate\Http\Response;
 
@@ -19,7 +19,19 @@ class GetCompleteTournamentController extends BasicController
     {
         return $this->responseBuilder
             ->setMessage(trans('endpoints.tournament.get_complete_tournament.ok'))
-            ->setResource('tournament', CompleteTournamentResource::make($tournament))
+            ->setResource('tournament', TournamentResource::make($tournament->load([
+                    'image',
+                    'tournamentType',
+                    'devices',
+                    'games',
+                    'tournamentFormat',
+                    'tournamentPrice',
+                    'tournamentPrizes',
+                    'tournamentRules',
+                    'tournamentSystems',
+                    'createdBy',
+                    'externalBracket',
+                ])))
             ->setStatusCode(Response::HTTP_OK)
             ->get();
     }
