+++ app/Http/Controllers/TournamentInscription/UpdateTournamentInscriptionController.php
@@ -6,7 +6,7 @@ use App\Contracts\Patterns\Builders\ResponseBuilder;
 use App\Events\TournamentInscriptionUpdated;
 use App\Http\Controllers\BasicController;
 use App\Http\Requests\UpdateTournamentInscriptionRequest;
-use App\Http\Resources\TournamentInscriptionResource;
+use App\Http\Resources\TournamentInscription\TournamentInscriptionResource;
 use App\Models\TournamentInscription;
 use App\Services\TournamentInscription\UpdateTournamentInscriptionService;
 use Illuminate\Http\Response;
